Extracted getAllServices and apply to k8sLib
